adds devcontainer configurations
With this configuration you can do choose a container when opening devcontainers. Most of the time you just want to jump into the php container.

But when you want to investigate something in the other container where the nextcloud code and logs are you can use the Switch Container command of vscode to switch to the other container.