smart tagging when scanning
you shouldn't just remove all tags and assign the one you acutally want to assign.

Instead this now checks for the existing tags of a file and only adds the one that actually should be added. And when the tag is alreaddfy assigned to the file it does nothing anymore.