Reduce number of activity events generated by tags (#111)
This introduces a new SystemTagObjectMapperWithoutActivity that uses a
dummy event system to avoid triggering activities.

Change TagService so that callers can decide if they want activities
generated for the tags or not. Change defaults to only trigger
activities for Malicious + Pup tags.

This still generates some events for the initial creation of the system
tags (done on first use), which is acceptable I think (it's only like 5
events).

Add a new test for the new behaviour.