adjusts debug settings to minimal settings
The client_host is autodiscovered by the HTTP-Headers like described on the [all_settings page](https://xdebug.org/docs/all_settings) of xdebug.
Search for client_discovery_header.

You only have to set a client_host for the cli client, but that cannot be done automatically because the container ip is rather dynamic (and depends on the docker-host configuration).